Added the required Rest API to the application.

+++ backend/src/main/java/com/prem/ta/repositories/FileRepository.java
@@ -11,8 +11,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface FileRepository
-        extends JpaRepository<File, Integer> {
+public interface FileRepository extends JpaRepository<File, Long> {
 
     Optional<File> findTopByTickerOrderByFileDateDesc(Ticker ticker);
